requirements.txt
python-telegram-bot==20.7
PyMuPDF==1.18.19
Pillow==8.4.0
apscheduler==3.7.0
Flask==2.1.2
python-telegram-bot==20.0
PyMuPDF==1.18.19
